requirements.txt
pyttsx3
openai
PyMuPDF
PyPDF2
python-dotenv
streamlit
speechrecognition
